requirements.txt
lightgbm
neptune-client
pandas
pillow
plotly
scikit-learn
shap
streamlit
tsfresh
xgboost
yaml